+++ lib/ServeAssets/getRoot.js
