+++ lib/ServeAssets/index.js
@@ -32,7 +32,7 @@ function ServeAssets() {
   // this.router.get('/normalize.css', this.getCssNormalize.bind(this));
 }
 
-ServeAssets.prototype.getRoot = require('./get-root');
+ServeAssets.prototype.getRoot = require('./getRoot');
 // ServeAssets.prototype.getJs = require('./get-js');
 // ServeAssets.prototype.getCss = require('./get-css');
 // ServeAssets.prototype.getCssReset = require('./get-css-reset');
